// backend/src/modules/media/events/media.event.types.ts
export enum MediaEventType {
  FILE_UPLOADED = 'media.file_uploaded',
  FILE_DELETED = 'media.file_deleted',

  PRODUCT_IMAGE_ATTACHED = 'media.product_image_attached',
  PRODUCT_IMAGE_UPDATED = 'media.product_image_updated',
  PRODUCT_IMAGE_DELETED = 'media.product_image_deleted',
  PRODUCT_IMAGE_PRIMARY_SET = 'media.product_image_primary_set',
  PRODUCT_IMAGES_REORDERED = 'media.product_images_reordered',

  BRAND_LOGO_UPDATED = 'media.brand_logo_updated',
  CATEGORY_IMAGE_UPDATED = 'media.category_image_updated',
  USER_AVATAR_UPDATED = 'media.user_avatar_updated',
  VARIANT_IMAGE_UPDATED = 'media.variant_image_updated',
}